fix(algoritmos): corrige ex4/ex5 e atualiza indice do README
- ex4: antecessor e sucessor estavam invertidos no return
- ex5: valor unitario lido como float para aceitar centavos
- README: linhas 4 e 5 apontam para os .py, progresso 5 de 16

+++ 1-semestre/algoritmos/exercicio1/ex5.py
@@ -1,6 +1,6 @@
 def compras():
     item = int(input("Entre a quantidade de itens: "))
-    valor = int(input("Entre o valor individual do item: "))
+    valor = float(input("Entre o valor individual do item: "))
     return item, valor
 
 def total(quant, preco):
